@@ -1,5 +1,11 @@
 package com.mtdevelopment.admin.domain.repository
 
+import com.mtdevelopment.admin.domain.model.OptimizedRouteWithOrders
+import com.mtdevelopment.core.model.Order
+
 interface GoogleRouteRepository {
-    suspend fun getOptimizedDeliveryPath(addresses: List<String>): List<Pair<Double, Double>>?
+    suspend fun getOptimizedDeliveryPath(
+        addresses: List<String>,
+        dailyOrders: List<Order>
+    ): OptimizedRouteWithOrders
 }
