@@ -1,17 +1,27 @@
 package com.mtdevelopment.admin.domain.usecase
 
+import com.mtdevelopment.admin.domain.model.OptimizedRouteWithOrders
+import com.mtdevelopment.admin.domain.repository.AdminDatastorePreference
 import com.mtdevelopment.admin.domain.repository.GoogleRouteRepository
+import com.mtdevelopment.core.model.Order
+import kotlinx.coroutines.flow.first
 
 class GetOptimizedDeliveryUseCase(
-    private val googleRouteRepository: GoogleRouteRepository
+    private val googleRouteRepository: GoogleRouteRepository,
+    private val adminDatastorePreference: AdminDatastorePreference
 ) {
 
+    // TODO: CLEANER MY DUDE
     suspend operator fun invoke(
-        addresses: List<String>
-    ): List<Pair<Double, Double>>? {
-        return googleRouteRepository.getOptimizedDeliveryPath(addresses)
-    }
+        addresses: List<String>,
+        dailyOrders: List<Order>
+    ): OptimizedRouteWithOrders {
 
-    // TODO: Keep in sharedPrefs to lower API use, and keep the order index with it
+        return if (adminDatastorePreference.dailyDeliveryPathGeocodedFlow.first() == null) {
+            googleRouteRepository.getOptimizedDeliveryPath(addresses, dailyOrders)
+        } else {
+            adminDatastorePreference.dailyDeliveryPathGeocodedFlow.first()!!
+        }
+    }
 
 }
