@@ -20,8 +20,10 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
@@ -30,6 +32,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleEventObserver
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.PermissionStatus
 import com.google.accompanist.permissions.isGranted
@@ -40,7 +44,7 @@ import com.google.accompanist.permissions.shouldShowRationale
 private enum class PermissionStep {
     INITIAL_PERMISSIONS,
     EXPLAIN_BACKGROUND_LOCATION,
-    REQUEST_BACKGROUND_LOCATION, // This step might directly lead to settings
+    REQUEST_BACKGROUND_LOCATION,
     CHECK_BATTERY_OPTIMIZATION,
     COMPLETED
 }
@@ -48,26 +52,22 @@ private enum class PermissionStep {
 @OptIn(ExperimentalPermissionsApi::class)
 @Composable
 fun RequestPermissionsScreen(
-    onPermissionsGrantedAndConfigured: () -> Unit, // All set, including potential battery optimization
-    onPermissionsProcessSkippedOrDenied: () -> Unit // User chose to skip or critical permissions denied
+    shouldShowOptimization: Boolean,
+    onPermissionsGrantedAndConfigured: () -> Unit,
+    onPermissionsProcessSkippedOrDenied: () -> Unit
 ) {
     val context = LocalContext.current
     var currentStep by remember { mutableStateOf(PermissionStep.INITIAL_PERMISSIONS) }
 
-    // Initial Permissions (Fine Location, Notifications)
     val initialPermissionsList = remember {
         listOfNotNull(
             Manifest.permission.ACCESS_FINE_LOCATION,
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) Manifest.permission.POST_NOTIFICATIONS else null
-            // FOREGROUND_SERVICE and FOREGROUND_SERVICE_LOCATION are manifest declared
-            // and their runtime effect for location is tied to ACCESS_FINE_LOCATION being granted
-            // and the service being started with the correct foregroundServiceType.
         ).distinct()
     }
     val initialPermissionsState =
         rememberMultiplePermissionsState(permissions = initialPermissionsList)
 
-    // Background Location Permission (Android Q+)
     val backgroundLocationPermissionString = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
         Manifest.permission.ACCESS_BACKGROUND_LOCATION
     } else null
@@ -76,30 +76,63 @@ fun RequestPermissionsScreen(
         rememberPermissionState(permission = it)
     }
 
-    // --- LaunchedEffects to manage state transitions ---
+    // State to trigger re-check of battery optimization on resume
+    var triggerBatteryRecheck by remember { mutableIntStateOf(0) }
+    val lifecycleOwner = androidx.lifecycle.compose.LocalLifecycleOwner.current
+
+    // Use DisposableEffect to manage lifecycle observer
+    DisposableEffect(lifecycleOwner, currentStep) {
+        val observer = LifecycleEventObserver { _, event ->
+            if (event == Lifecycle.Event.ON_RESUME) {
+                Log.d("RequestPermissionsScreen", "ON_RESUME triggered. Current step: $currentStep")
+                if (currentStep == PermissionStep.CHECK_BATTERY_OPTIMIZATION) {
+                    Log.d(
+                        "RequestPermissionsScreen",
+                        "ON_RESUME: Triggering battery optimization re-check."
+                    )
+                    triggerBatteryRecheck++ // Increment to trigger recomposition for the remember block
+                }
+            }
+        }
+        lifecycleOwner.lifecycle.addObserver(observer)
+        onDispose {
+            Log.d(
+                "RequestPermissionsScreen",
+                "ON_DISPOSE: Removing lifecycle observer for step $currentStep"
+            )
+            lifecycleOwner.lifecycle.removeObserver(observer)
+        }
+    }
 
     LaunchedEffect(initialPermissionsState.allPermissionsGranted) {
         if (initialPermissionsState.allPermissionsGranted) {
-            if (currentStep == PermissionStep.INITIAL_PERMISSIONS) { // Ensure this transition happens once
+            if (currentStep == PermissionStep.INITIAL_PERMISSIONS) {
                 currentStep =
                     if (backgroundLocationPermissionString != null && backgroundLocationPermissionState?.status?.isGranted == false) {
                         PermissionStep.EXPLAIN_BACKGROUND_LOCATION
                     } else {
                         PermissionStep.CHECK_BATTERY_OPTIMIZATION
                     }
+                Log.d(
+                    "RequestPermissionsScreen",
+                    "Initial permissions granted. New step: $currentStep"
+                )
             }
         }
     }
 
     LaunchedEffect(backgroundLocationPermissionState?.status) {
-        if (backgroundLocationPermissionState?.status?.isGranted == true) {
+        val bgStatus = backgroundLocationPermissionState?.status
+        if (bgStatus?.isGranted == true) {
             if (currentStep == PermissionStep.REQUEST_BACKGROUND_LOCATION || currentStep == PermissionStep.EXPLAIN_BACKGROUND_LOCATION) {
                 currentStep = PermissionStep.CHECK_BATTERY_OPTIMIZATION
+                Log.d(
+                    "RequestPermissionsScreen",
+                    "Background location granted. New step: $currentStep"
+                )
             }
         }
     }
-
-    // --- UI based on currentStep ---
 
     Column(
         modifier = Modifier
@@ -134,26 +167,29 @@ fun RequestPermissionsScreen(
 
             PermissionStep.EXPLAIN_BACKGROUND_LOCATION -> {
                 Text(
-                    "Pour un suivi continu même lorsque l'application est en arrière-plan, veuillez autoriser la localisation \"Toujours\" ou \"Seulement si l'appli est en cours d'utilisation\" dans l'écran suivant.",
+                    "Pour un suivi continu même lorsque l'application est en arrière-plan, veuillez autoriser la localisation \"Toujours\" dans l'écran suivant.",
                     textAlign = TextAlign.Center, style = MaterialTheme.typography.bodyLarge
                 )
                 Spacer(modifier = Modifier.height(16.dp))
                 Button(onClick = {
                     currentStep = PermissionStep.REQUEST_BACKGROUND_LOCATION
-                    // Small delay to ensure state change before launching, if needed, though usually direct is fine.
                     backgroundLocationPermissionState?.launchPermissionRequest()
                 }) {
                     Text("Compris, ouvrir les paramètres")
                 }
                 Spacer(modifier = Modifier.height(8.dp))
-                TextButton(onClick = { currentStep = PermissionStep.CHECK_BATTERY_OPTIMIZATION }) {
+                TextButton(onClick = {
+                    currentStep = PermissionStep.CHECK_BATTERY_OPTIMIZATION
+                    Log.d(
+                        "RequestPermissionsScreen",
+                        "Skipping background location. New step: $currentStep"
+                    )
+                }) {
                     Text("Continuer sans localisation en arrière-plan")
                 }
             }
 
             PermissionStep.REQUEST_BACKGROUND_LOCATION -> {
-                // This state is mostly transient as launchPermissionRequest often navigates away.
-                // If the user comes back without granting:
                 if (backgroundLocationPermissionState?.status?.isGranted == false) {
                     Text(
                         "La localisation en arrière-plan n'a pas été accordée. Vous pouvez l'activer manuellement dans les paramètres de l'application pour un meilleur suivi.",
@@ -166,12 +202,25 @@ fun RequestPermissionsScreen(
                     Spacer(modifier = Modifier.height(8.dp))
                     TextButton(onClick = {
                         currentStep = PermissionStep.CHECK_BATTERY_OPTIMIZATION
+                        Log.d(
+                            "RequestPermissionsScreen",
+                            "Continuing without background location after failed attempt. New step: $currentStep"
+                        )
                     }) {
                         Text("Continuer quand même")
                     }
                 } else if (backgroundLocationPermissionState?.status?.isGranted == true) {
-                    // Should be caught by LaunchedEffect, but as a safeguard
-                    LaunchedEffect(Unit) { currentStep = PermissionStep.CHECK_BATTERY_OPTIMIZATION }
+                    // This case should ideally be caught by the LaunchedEffect on backgroundLocationPermissionState.status
+                    // If somehow reached, ensure transition.
+                    LaunchedEffect(Unit) {
+                        if (currentStep != PermissionStep.CHECK_BATTERY_OPTIMIZATION && currentStep != PermissionStep.COMPLETED) {
+                            currentStep = PermissionStep.CHECK_BATTERY_OPTIMIZATION
+                            Log.d(
+                                "RequestPermissionsScreen",
+                                "Background location detected as granted in REQUEST_BACKGROUND_LOCATION step (safeguard). New step: $currentStep"
+                            )
+                        }
+                    }
                 } else {
                     Text(
                         "Redirection vers les paramètres de localisation en arrière-plan...",
@@ -181,56 +230,116 @@ fun RequestPermissionsScreen(
             }
 
             PermissionStep.CHECK_BATTERY_OPTIMIZATION -> {
-                val powerManager = context.getSystemService(Context.POWER_SERVICE) as PowerManager
-                val isIgnoringOptimizations = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                    powerManager.isIgnoringBatteryOptimizations(context.packageName)
-                } else {
-                    true // Not applicable before Marshmallow
-                }
+                if (shouldShowOptimization) {
+                    val isIgnoringOptimizations = remember(triggerBatteryRecheck) {
+                        val pm = context.getSystemService(Context.POWER_SERVICE) as PowerManager
+                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                            Log.d(
+                                "RequestPermissionsScreen",
+                                "Checking battery optimization status. Trigger count: $triggerBatteryRecheck. Package: ${context.packageName}"
+                            )
+                            val status = pm.isIgnoringBatteryOptimizations(context.packageName)
+                            Log.d(
+                                "RequestPermissionsScreen",
+                                "pm.isIgnoringBatteryOptimizations returned: $status for package ${context.packageName}"
+                            )
+                            status
+                        } else {
+                            true // Not applicable before Marshmallow
+                        }
+                    }
 
-                if (!isIgnoringOptimizations && Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                    Text(
-                        "Pour assurer la fiabilité du suivi de livraison, il est recommandé de désactiver l'optimisation de la batterie pour cette application.",
-                        textAlign = TextAlign.Center, style = MaterialTheme.typography.bodyLarge
-                    )
-                    Spacer(modifier = Modifier.height(16.dp))
-                    Button(onClick = {
-                        requestIgnoreBatteryOptimizations(context)
-                        // After this, we assume the user handled it or we proceed anyway.
-                        currentStep = PermissionStep.COMPLETED
-                        onPermissionsGrantedAndConfigured()
-                    }) {
-                        Text("Ouvrir les paramètres d'optimisation")
+                    // Log the evaluated status outside the remember block to see it on each recomposition of this step
+                    LaunchedEffect(isIgnoringOptimizations) {
+                        Log.d(
+                            "RequestPermissionsScreen",
+                            "CHECK_BATTERY_OPTIMIZATION (recomposition): isIgnoringOptimizations = $isIgnoringOptimizations"
+                        )
                     }
-                    Spacer(modifier = Modifier.height(8.dp))
-                    TextButton(onClick = {
-                        currentStep = PermissionStep.COMPLETED
-                        onPermissionsGrantedAndConfigured() // Proceed even if they don't optimize
-                    }) {
-                        Text("Ne pas optimiser pour l'instant")
+
+                    if (!isIgnoringOptimizations && Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                        Text(
+                            "Pour assurer la fiabilité du suivi de livraison, il est recommandé de désactiver l'optimisation de la batterie pour cette application.",
+                            textAlign = TextAlign.Center, style = MaterialTheme.typography.bodyLarge
+                        )
+                        Spacer(modifier = Modifier.height(16.dp))
+                        Button(onClick = {
+                            requestIgnoreBatteryOptimizations(context)
+                            Log.d(
+                                "RequestPermissionsScreen",
+                                "Requested to ignore battery optimizations. Waiting for resume to re-check."
+                            )
+                        }) {
+                            Text("Ouvrir les paramètres d'optimisation")
+                        }
+                        Spacer(modifier = Modifier.height(8.dp))
+                        TextButton(onClick = {
+                            Log.d(
+                                "RequestPermissionsScreen",
+                                "User chose not to optimize battery. Proceeding to COMPLETED."
+                            )
+                            currentStep =
+                                PermissionStep.COMPLETED // This will trigger COMPLETED step logic
+                        }) {
+                            Text("Ne pas optimiser pour l'instant")
+                        }
+                    } else { // Battery optimization is already ignored or not applicable
+                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                            Text(
+                                "Vérification de l'optimisation de la batterie : OK.",
+                                textAlign = TextAlign.Center,
+                                style = MaterialTheme.typography.bodyMedium
+                            )
+                            Spacer(modifier = Modifier.height(16.dp))
+                        }
+                        // Transition to COMPLETED. The callback will be handled there.
+                        // Key with currentStep to ensure this runs if we enter this 'else' branch and are not yet COMPLETED.
+                        LaunchedEffect(currentStep) {
+                            if (currentStep != PermissionStep.COMPLETED) {
+                                Log.d(
+                                    "RequestPermissionsScreen",
+                                    "Battery optimization OK or not needed. Setting step to COMPLETED."
+                                )
+                                currentStep = PermissionStep.COMPLETED
+                            }
+                        }
                     }
                 } else {
-                    // Already ignoring or not applicable
-                    LaunchedEffect(Unit) {
-                        currentStep = PermissionStep.COMPLETED
-                        onPermissionsGrantedAndConfigured()
-                    }
+                    currentStep = PermissionStep.COMPLETED
+                    Log.d(
+                        "RequestPermissionsScreen",
+                        "Battery optimization not needed. Setting step to COMPLETED."
+                    )
                 }
             }
 
             PermissionStep.COMPLETED -> {
-                // This state means the flow is done. The parent composable should typically
-                // stop showing RequestPermissionsScreen once onPermissionsGrantedAndConfigured or
-                // onPermissionsProcessSkippedOrDenied is called.
-                // If shown, it means the parent didn't react.
                 Text("Configuration des permissions terminée.", textAlign = TextAlign.Center)
-                LaunchedEffect(Unit) {
-                    // Ensure callback is triggered if somehow stuck here
-                    if (initialPermissionsState.allPermissionsGranted &&
-                        (backgroundLocationPermissionString == null || backgroundLocationPermissionState?.status?.isGranted == true)
-                    ) {
+                // This LaunchedEffect will run once when currentStep becomes COMPLETED,
+                // or if permission states change while in COMPLETED state.
+                LaunchedEffect(
+                    initialPermissionsState.allPermissionsGranted,
+                    backgroundLocationPermissionState?.status // This key can be null
+                ) {
+                    Log.d(
+                        "RequestPermissionsScreen",
+                        "In COMPLETED step. Evaluating final callback. Initial granted: ${initialPermissionsState.allPermissionsGranted}, BG status: ${backgroundLocationPermissionState?.status}"
+                    )
+
+                    val bgPermissionGrantedOrNotApplicable =
+                        backgroundLocationPermissionString == null || backgroundLocationPermissionState?.status?.isGranted == true
+
+                    if (initialPermissionsState.allPermissionsGranted && bgPermissionGrantedOrNotApplicable) {
+                        Log.d(
+                            "RequestPermissionsScreen",
+                            "COMPLETED: Calling onPermissionsGrantedAndConfigured."
+                        )
                         onPermissionsGrantedAndConfigured()
                     } else {
+                        Log.d(
+                            "RequestPermissionsScreen",
+                            "COMPLETED: Calling onPermissionsProcessSkippedOrDenied due to permission state. Initial: ${initialPermissionsState.allPermissionsGranted}, BG relevant: ${backgroundLocationPermissionString != null}, BG granted: ${backgroundLocationPermissionState?.status?.isGranted}"
+                        )
                         onPermissionsProcessSkippedOrDenied()
                     }
                 }
@@ -238,19 +347,23 @@ fun RequestPermissionsScreen(
         }
     }
 
-    // Initial request for non-background permissions when the composable first appears
-    // and currentStep is INITIAL_PERMISSIONS
-    LaunchedEffect(currentStep, initialPermissionsState.allPermissionsGranted) {
-        if (currentStep == PermissionStep.INITIAL_PERMISSIONS && !initialPermissionsState.allPermissionsGranted) {
-            // Check if we should show rationale before the first launch,
-            // though Accompanist handles this well.
-            // This ensures the dialog appears on first display if permissions are missing.
+    // Initial request for non-background permissions
+    LaunchedEffect(
+        initialPermissionsState.allPermissionsGranted,
+        currentStep
+    ) { // Added currentStep as key
+        if (currentStep == PermissionStep.INITIAL_PERMISSIONS && !initialPermissionsState.allPermissionsGranted &&
+            !initialPermissionsState.permissions.any { it.status.shouldShowRationale } // More precise rationale check
+        ) {
+            Log.d(
+                "RequestPermissionsScreen",
+                "Initial launch of permission request for initial permissions."
+            )
             initialPermissionsState.launchMultiplePermissionRequest()
         }
     }
 }
 
-// Helper to open app-specific settings
 private fun openAppSettings(context: Context) {
     val intent = Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS).apply {
         data = Uri.fromParts("package", context.packageName, null)
@@ -263,7 +376,6 @@ private fun openAppSettings(context: Context) {
     }
 }
 
-// Helper to request ignoring battery optimizations
 @SuppressLint("BatteryLife")
 private fun requestIgnoreBatteryOptimizations(context: Context) {
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
@@ -276,8 +388,7 @@ private fun requestIgnoreBatteryOptimizations(context: Context) {
             context.startActivity(intent)
         } catch (e: Exception) {
             Log.e("BatteryOpt", "Could not open battery optimization settings", e)
-            // As a fallback, open general app settings if specific intent fails
-            openAppSettings(context)
+            openAppSettings(context) // Fallback
         }
     }
 }
